Move logic to submodule
